Backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
django-extensions==3.2.3
django-graphql-jwt==0.4.0
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
graphene-django==3.2.2
Werkzeug==3.0.3
watchdog==4.0.1
psycopg2==2.9.9
pathlib2==2.3.7.post1
pgsql==2.2